# bytes fingerprints which speeds up decoding 100x