--- a/Cargo.toml
+++ b/Cargo.toml
@@ -32,6 +32,7 @@ libm = { version = "0.2.1", optional = true }
 serde = { version = "1.0.104", optional = true, features = ["derive"] }
 num-traits = { version = "0.2.12", optional = true }
 rand = { version = "0.8", optional = true }
+bytes = "1.1"
 
 [dev-dependencies]
 criterion = "0.3.0"
@@ -61,6 +62,7 @@ harness = false
 required-features = ["serde"]
 
 [features]
-default = ["uniform-random", "binary-fuse"]
+default = ["uniform-random", "binary-fuse", "serde"]
+serde = ["dep:serde", "bytes/serde"]
 uniform-random = ["rand"]
 binary-fuse = ["libm"]