```
custom/serialize        time:   [4.5760 us 4.6046 us 4.6413 us]
                        change: [+6.6152% +8.1193% +9.5426%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
custom/deserialize-from-slice
                        time:   [2.8123 us 2.8281 us 2.8507 us]
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
custom/deserialize-from-bytes
                        time:   [27.960 ns 28.113 ns 28.291 ns]
Found 15 outliers among 100 measurements (15.00%)
  2 (2.00%) high mild
  13 (13.00%) high severe

bincode/serialize       time:   [3.8784 us 3.8886 us 3.9011 us]
                        change: [-96.096% -96.076% -96.059%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low severe
  4 (4.00%) high mild
  5 (5.00%) high severe
bincode/deserialize     time:   [3.7303 us 3.7468 us 3.7656 us]
                        change: [-98.287% -98.281% -98.275%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  6 (6.00%) low mild
  2 (2.00%) high mild
  8 (8.00%) high severe
```

Signed-off-by: youjiali1995 <zlwgx1023@gmail.com>